doc: enforce code styling

+++ code/mobile/android/PhoneVR/app/src/main/java/viritualisres/phonevr/SettingsActivity.kt
@@ -1,10 +1,10 @@
+/* (C)2023 */
 package viritualisres.phonevr
 
 import android.annotation.SuppressLint
 import android.content.Context
 import android.content.DialogInterface
 import android.content.Intent
-import android.net.Uri
 import android.os.Bundle
 import android.text.method.ScrollingMovementMethod
 import android.util.Log
@@ -15,12 +15,11 @@ import android.widget.TextView
 import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.content.FileProvider
-import kotlinx.android.synthetic.main.activity_settings.*
 import java.io.*
 import java.util.*
 import java.util.zip.ZipEntry
 import java.util.zip.ZipOutputStream
-
+import kotlinx.android.synthetic.main.activity_settings.*
 
 class SettingsActivity : AppCompatActivity() {
 
@@ -37,15 +36,15 @@ class SettingsActivity : AppCompatActivity() {
         etLogView.setMovementMethod(ScrollingMovementMethod())
         etLogView.setFocusable(false)
         val touchListener = OnTouchListener { v, motionEvent ->
-                v.parent.requestDisallowInterceptTouchEvent(true)
-                when (motionEvent.action and MotionEvent.ACTION_MASK) {
-                    MotionEvent.ACTION_UP -> v.parent.requestDisallowInterceptTouchEvent(false)
-                }
+            v.parent.requestDisallowInterceptTouchEvent(true)
+            when (motionEvent.action and MotionEvent.ACTION_MASK) {
+                MotionEvent.ACTION_UP -> v.parent.requestDisallowInterceptTouchEvent(false)
+            }
             false
         }
         etLogView.setOnTouchListener(touchListener)
 
-        val btShareLogs: Button  = findViewById(R.id.btShareLogs)
+        val btShareLogs: Button = findViewById(R.id.btShareLogs)
         btShareLogs.setOnClickListener {
             val intentShareFile = Intent(Intent.ACTION_SEND)
 
@@ -57,42 +56,48 @@ class SettingsActivity : AppCompatActivity() {
 
             if (fileWithinMyDir.exists()) {
                 if (fileWithinMyDir.listFiles()?.size!! > 1) {
-                    if(fileZip.exists())
-                        fileZip.delete()
+                    if (fileZip.exists()) fileZip.delete()
 
-                    zipFileAtPath( pvrFilesFolder, pvrZipFilesFolder)
+                    zipFileAtPath(pvrFilesFolder, pvrZipFilesFolder)
                     intentShareFile.type = "application/zip"
-                    intentShareFile.putExtra(Intent.EXTRA_STREAM,
-                            FileProvider.getUriForFile(Objects.requireNonNull(applicationContext),
-                                    BuildConfig.APPLICATION_ID + ".provider",
-                                    fileZip))
-                }
-                else {
+                    intentShareFile.putExtra(
+                        Intent.EXTRA_STREAM,
+                        FileProvider.getUriForFile(
+                            Objects.requireNonNull(applicationContext),
+                            BuildConfig.APPLICATION_ID + ".provider",
+                            fileZip))
+                } else {
                     intentShareFile.type = "application/txt"
-                    intentShareFile.putExtra(Intent.EXTRA_STREAM,
-                            FileProvider.getUriForFile(Objects.requireNonNull(applicationContext),
-                                    BuildConfig.APPLICATION_ID + ".provider",
-                                    File(getExternalFilesDir(null).toString() + "/PVR/pvrlog.txt")))
+                    intentShareFile.putExtra(
+                        Intent.EXTRA_STREAM,
+                        FileProvider.getUriForFile(
+                            Objects.requireNonNull(applicationContext),
+                            BuildConfig.APPLICATION_ID + ".provider",
+                            File(getExternalFilesDir(null).toString() + "/PVR/pvrlog.txt")))
                 }
-                intentShareFile.putExtra(Intent.EXTRA_SUBJECT,
-                        "Sharing pvrlog File(s)...")
+                intentShareFile.putExtra(Intent.EXTRA_SUBJECT, "Sharing pvrlog File(s)...")
                 intentShareFile.putExtra(Intent.EXTRA_TEXT, "Sharing pvrlog File(s)...")
                 startActivity(Intent.createChooser(intentShareFile, "Share File"))
             }
         }
 
-        val btOpenLogs: Button  = findViewById(R.id.btOpenLogs)
+        val btOpenLogs: Button = findViewById(R.id.btOpenLogs)
         btOpenLogs.setOnClickListener {
             val dir = File(getExternalFilesDir(null).toString() + "/PVR/")
             if (dir.listFiles()?.size!! > 1) {
                 val items = arrayOfNulls<CharSequence>(2)
-                items[0] = "pvrlog.txt"; items[1] = "pvrDebuglog.txt"
+                items[0] = "pvrlog.txt"
+                items[1] = "pvrDebuglog.txt"
 
                 val builder = AlertDialog.Builder(this@SettingsActivity)
-                builder.setTitle(R.string.pick_logfile)
-                        .setItems(R.array.logsFiles, DialogInterface.OnClickListener { _, which ->
+                builder
+                    .setTitle(R.string.pick_logfile)
+                    .setItems(
+                        R.array.logsFiles,
+                        DialogInterface.OnClickListener { _, which ->
                             Log.d("PVR_JAVA", "Which $which")
-                            OpenFileWithIntent(File(getExternalFilesDir(null).toString() + "/PVR/" + items[which]))
+                            OpenFileWithIntent(
+                                File(getExternalFilesDir(null).toString() + "/PVR/" + items[which]))
                         })
                 builder.create()
                 builder.show()
@@ -102,17 +107,16 @@ class SettingsActivity : AppCompatActivity() {
         }
     }
     /*
-    * Zips a file at a location and places the resulting zip file at the toLocation
-    * Example: zipFileAtPath("downloads/myfolder", "downloads/myFolder.zip");
-    */
+     * Zips a file at a location and places the resulting zip file at the toLocation
+     * Example: zipFileAtPath("downloads/myfolder", "downloads/myFolder.zip");
+     */
     private fun zipFileAtPath(sourcePath: String, toLocation: String?): Boolean {
         val BUFFER = 2048
         val sourceFile = File(sourcePath)
         try {
             var origin: BufferedInputStream? = null
             val dest = FileOutputStream(toLocation)
-            val out = ZipOutputStream(BufferedOutputStream(
-                    dest))
+            val out = ZipOutputStream(BufferedOutputStream(dest))
             if (sourceFile.isDirectory) {
                 zipSubFolder(out, sourceFile, sourceFile.parent.length)
             } else {
@@ -120,7 +124,8 @@ class SettingsActivity : AppCompatActivity() {
                 val fi = FileInputStream(sourcePath)
                 origin = BufferedInputStream(fi, BUFFER)
                 val entry = ZipEntry(getLastPathComponent(sourcePath))
-                entry.setTime(sourceFile.lastModified()) // to keep modification time after unzipping
+                entry.setTime(
+                    sourceFile.lastModified()) // to keep modification time after unzipping
                 out.putNextEntry(entry)
                 var count: Int = 0
                 while (origin.read(data, 0, BUFFER).also({ count = it }) != -1) {
@@ -136,11 +141,10 @@ class SettingsActivity : AppCompatActivity() {
     }
 
     /*
-    * Zips a subfolder
-    */
+     * Zips a subfolder
+     */
     @Throws(IOException::class)
-    private fun zipSubFolder(out: ZipOutputStream, folder: File,
-                             basePathLength: Int) {
+    private fun zipSubFolder(out: ZipOutputStream, folder: File, basePathLength: Int) {
         val BUFFER = 2048
         val fileList = folder.listFiles()
         var origin: BufferedInputStream? = null
@@ -150,8 +154,7 @@ class SettingsActivity : AppCompatActivity() {
             } else {
                 val data = ByteArray(BUFFER)
                 val unmodifiedFilePath = file.path
-                val relativePath = unmodifiedFilePath
-                        .substring(basePathLength)
+                val relativePath = unmodifiedFilePath.substring(basePathLength)
                 val fi = FileInputStream(unmodifiedFilePath)
                 origin = BufferedInputStream(fi, BUFFER)
                 val entry = ZipEntry(relativePath)
@@ -167,20 +170,23 @@ class SettingsActivity : AppCompatActivity() {
     }
 
     /*
-    * gets the last path component
-    *
-    * Example: getLastPathComponent("downloads/example/fileToZip");
-    * Result: "fileToZip"
-    */
+     * gets the last path component
+     *
+     * Example: getLastPathComponent("downloads/example/fileToZip");
+     * Result: "fileToZip"
+     */
     private fun getLastPathComponent(filePath: String): String? {
         val segments = filePath.split("/".toRegex()).toTypedArray()
         return if (segments.size == 0) "" else segments[segments.size - 1]
     }
 
-    private fun OpenFileWithIntent(file: File)
-    {
+    private fun OpenFileWithIntent(file: File) {
         // Get URI and MIME type of file
-        val uri = FileProvider.getUriForFile(Objects.requireNonNull(applicationContext), BuildConfig.APPLICATION_ID + ".provider", file)
+        val uri =
+            FileProvider.getUriForFile(
+                Objects.requireNonNull(applicationContext),
+                BuildConfig.APPLICATION_ID + ".provider",
+                file)
         val mime = contentResolver.getType(uri)
 
         // Open file with user selected app
@@ -191,7 +197,7 @@ class SettingsActivity : AppCompatActivity() {
         startActivity(intent)
     }
 
-    private fun setTextToLogger(et: TextView){
+    private fun setTextToLogger(et: TextView) {
         val file: File = File(getExternalFilesDir(null).toString() + "/PVR/pvrlog.txt")
         et.setText(tail2(file, 100))
     }
@@ -229,10 +235,10 @@ class SettingsActivity : AppCompatActivity() {
             e.printStackTrace()
             ""
         } finally {
-            if (fileHandler != null) try {
-                fileHandler.close()
-            } catch (e: IOException) {
-            }
+            if (fileHandler != null)
+                try {
+                    fileHandler.close()
+                } catch (e: IOException) {}
         }
     }
 
@@ -247,9 +253,12 @@ class SettingsActivity : AppCompatActivity() {
             val edit = prefs.edit()
             with(edit) {
                 putString(pcIpKey, pcIp.text.toString())
-                if(Util_IsVaildPort(connPort.text.toString().toInt())) putInt(connPortKey, connPort.text.toString().toInt())
-                if(Util_IsVaildPort(videoPort.text.toString().toInt())) putInt(videoPortKey, videoPort.text.toString().toInt())
-                if(Util_IsVaildPort(posePort.text.toString().toInt())) putInt(posePortKey, posePort.text.toString().toInt())
+                if (Util_IsVaildPort(connPort.text.toString().toInt()))
+                    putInt(connPortKey, connPort.text.toString().toInt())
+                if (Util_IsVaildPort(videoPort.text.toString().toInt()))
+                    putInt(videoPortKey, videoPort.text.toString().toInt())
+                if (Util_IsVaildPort(posePort.text.toString().toInt()))
+                    putInt(posePortKey, posePort.text.toString().toInt())
                 putInt(resMulKey, resMul.text.toString().toInt())
                 putFloat(mt2phKey, mt2ph.text.toString().replace(',', '.').toFloat())
                 putFloat(offFovKey, offFov.text.toString().replace(',', '.').toFloat())
@@ -257,11 +266,9 @@ class SettingsActivity : AppCompatActivity() {
                 putBoolean(debugKey, debug.isChecked)
                 apply()
             }
-        }
-        catch (e: Exception)
-        {
-            Log.d("PVR-Java", "Exception caught in savePrefs.SettingsActivity : " + e.message);
-            e.printStackTrace();
+        } catch (e: Exception) {
+            Log.d("PVR-Java", "Exception caught in savePrefs.SettingsActivity : " + e.message)
+            e.printStackTrace()
         }
     }
 
@@ -283,8 +290,7 @@ class SettingsActivity : AppCompatActivity() {
         debug.isChecked = prefs.getBoolean(debugKey, debugDef)
     }
 
-    private fun Util_IsVaildPort(port: Int) : Boolean
-    {
-        return ((port > 0)  && (port <= 65536));
+    private fun Util_IsVaildPort(port: Int): Boolean {
+        return ((port > 0) && (port <= 65536))
     }
 }
